chore(evals): cleanup batch — relocate poc/, flatten metrics-panel, wire aggregation, misc
Relocate poc/* and metrics-panel/classic/* out; delete unused scroll-edge-fades;
drop vestigial includeLabel; revert store persist name; wire the aggregation
dropdown into the score cards; fix resize-listener leak; collapse groups CTE;
hide data/target/metadata columns by default; trim comments.
