+++ frontend/components/evaluation/metrics-panel/binary-card.tsx
