+++ frontend/components/evaluation/row-score-chips/use-hover-grow-card.ts
